feat(join): cross join - refactor collection
