--- lib/array_collection/collect_components/mapping.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-module ArrayCollection
-  module CollectComponents
-    # Array mapping
-    module Mapping
-      def map(&block)
-        clone(ArrayCollection::CollectionArray.map(@items, &block))
-      end
-    end
-  end
-end
